pyproject.toml
[project]
name = "ksaitex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "markdown-it-py>=4.0.0",
    "typer>=0.21.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
